requirements.txt
agora_token_builder==1.0.0
anyio==3.7.0
boto3==1.28.20
boto3-stubs==1.28.20
botocore==1.31.20
botocore-stubs==1.31.30
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.2
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.97.0
fastapi-pagination==0.12.12
firebase-admin==6.2.0
gcloud==0.18.3
geographiclib==2.0
geopy==2.4.0
google-api-core==2.11.1
google-api-python-client==2.94.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.2
grpcio-status==1.56.2
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httptools==0.5.0
httpx==0.27.0
idna==3.4
importlib-metadata==6.6.0
jmespath==1.0.1
jwcrypto==1.5.6
mongo==0.2.0
motor==3.1.2
msgpack==1.0.5
mypy-boto3-rekognition==1.28.21
mypy-boto3-s3==1.28.27
mypy-boto3-s3control==1.28.16
mypy-boto3-s3outposts==1.28.16
oauth2client==4.1.3
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==1.10.9
PyJWT==2.8.0
pymongo==4.3.3
pyparsing==3.1.0
Pyrebase4==4.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jose==3.3.0
python-jwt==4.1.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
types-awscrt==0.19.0
types-s3transfer==0.6.2
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==1.26.16
uuid==1.30
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
zipp==3.15.0
